requirements.txt
pandas==3.0.2
psycopg-binary==3.3.3
python-dotenv==1.2.2
requests==2.33.1
streamlit==1.56.0
psycopg
